Changes on support for failure listener. Now abstract batch implements the error handling case by default.
