+++ src/main/java/com/feedzai/commons/sql/abstraction/engine/AbstractDatabaseEngine.java
@@ -15,10 +15,9 @@
  */
 package com.feedzai.commons.sql.abstraction.engine;
 
-import com.feedzai.commons.sql.abstraction.OnFailureListener;
+import com.feedzai.commons.sql.abstraction.FailureListener;
 import com.feedzai.commons.sql.abstraction.batch.AbstractBatch;
 import com.feedzai.commons.sql.abstraction.batch.DefaultBatch;
-import com.feedzai.commons.sql.abstraction.batch.FailureHandlerBatch;
 import com.feedzai.commons.sql.abstraction.ddl.*;
 import com.feedzai.commons.sql.abstraction.dml.Expression;
 import com.feedzai.commons.sql.abstraction.dml.dialect.Dialect;
@@ -798,22 +797,12 @@ public abstract class AbstractDatabaseEngine implements DatabaseEngine {
 
     @Override
     public AbstractBatch createBatch(final int batchSize, final long batchTimeout, final String batchName) {
-        return DefaultBatch.create(this, batchName, batchSize, batchTimeout, properties.getMaximumAwaitTimeBatchShutdown());
+        return createBatch(batchSize, batchTimeout, batchName, Optional.empty());
     }
 
     @Override
-    public AbstractBatch createBatch(int batchSize, long batchTimeout, String batchName, OnFailureListener failureListener) {
-        if (failureListener == null) {
-            return createBatch(batchSize, batchTimeout, batchName);
-        } else {
-            return FailureHandlerBatch.create(
-                    this,
-                    batchName,
-                    batchSize,
-                    batchTimeout,
-                    properties.getMaximumAwaitTimeBatchShutdown(), failureListener
-            );
-        }
+    public AbstractBatch createBatch(int batchSize, long batchTimeout, String batchName, final Optional<FailureListener> failureListener) {
+        return DefaultBatch.create(this, batchName, batchSize, batchTimeout, properties.getMaximumAwaitTimeBatchShutdown(), failureListener);
     }
 
     /**
