+++ src/main/java/com/feedzai/commons/sql/abstraction/engine/DatabaseEngine.java
@@ -15,7 +15,7 @@
  */
 package com.feedzai.commons.sql.abstraction.engine;
 
-import com.feedzai.commons.sql.abstraction.OnFailureListener;
+import com.feedzai.commons.sql.abstraction.FailureListener;
 import com.feedzai.commons.sql.abstraction.batch.AbstractBatch;
 import com.feedzai.commons.sql.abstraction.ddl.DbColumnType;
 import com.feedzai.commons.sql.abstraction.ddl.DbEntity;
@@ -31,6 +31,7 @@ import com.feedzai.commons.sql.abstraction.entry.EntityEntry;
 import java.sql.Connection;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Properties;
 
 /**
@@ -247,6 +248,9 @@ public interface DatabaseEngine {
     /**
      * Creates a new batch that periodically flushes a batch. A flush will also occur when the maximum number of
      * statements in the batch is reached.
+     * <p>
+     * If desired you may provide an {@link Optional} {@link FailureListener} that will be invoked when the batch
+     * fails to persist data.
      * <p/>
      * Please be sure to call {@link AbstractBatch#destroy() } before closing the session with the database.
      *
@@ -256,7 +260,7 @@ public interface DatabaseEngine {
      * @param failureListener Failure listener to execute custom behavior when the batch fails to persist. It may be null.
      * @return The batch.
      */
-    AbstractBatch createBatch(final int batchSize, final long batchTimeout, final String batchName, final OnFailureListener failureListener);
+    AbstractBatch createBatch(final int batchSize, final long batchTimeout, final String batchName, final Optional<FailureListener> failureListener);
 
     /**
      * Checks if the connection is alive.
